+++ R/tt_download_file.R
@@ -17,10 +17,7 @@
 #'
 #' agencies <- tt_download_file(tt_gh, 1)
 #' launches <- tt_download_file(tt_gh, "launches.csv")
-tt_download_file <- function(tt,
-                             x,
-                             ...,
-                             auth = gh::gh_token()) {
+tt_download_file <- function(tt, x, ..., auth = gh::gh_token()) {
   file_info <- attr(tt, ".files")
   tt_date <- attr(tt, ".date")
   call <- rlang::caller_env()
@@ -58,7 +55,8 @@ tt_download_file_raw <- function(tt_date, target) {
 }
 
 tt_parse_download <- function(gh_response, ..., data_type, delim = NA) {
-  switch(data_type,
+  switch(
+    data_type,
     "rds" = return(tt_parse_rds(gh_response, ...)), # nocov 3 examples
     "xls" = return(tt_parse_excel(gh_response, ...)), # nocov 0 examples
     "xlsx" = return(tt_parse_excel(gh_response, ...)),
@@ -87,7 +85,8 @@ tt_guess_delim <- function(delim, data_type) {
   if (!is.na(delim)) {
     return(delim)
   }
-  switch(tolower(data_type),
+  switch(
+    tolower(data_type),
     "csv" = ",", # nocov
     "tsv" = "\t",
     "," # nocov